+++ formerbox/data/indexed_dataset.py
@@ -1,4 +1,5 @@
 import inspect
+import logging
 import os
 import struct
 from abc import ABCMeta, abstractmethod
@@ -13,6 +14,9 @@ from formerbox.utils import all_subclasses, path_to_posix
 from numpy import float32, float64, int8, int16, int32, int64, uint8, uint16
 from torch.utils.data import Dataset
 from typing_extensions import Protocol
+
+logger = logging.getLogger(__name__)
+
 
 element_codes = {
     1: uint8,