+++ formerbox/tasks/code/code_tokenization.py
@@ -1,12 +1,15 @@
+import logging
 from typing import Any, List, Optional, Text, Union
 
-from formerbox.tasks import TransformerTokenizerFast
+from formerbox.tasks import ByteLevelBPETokenizerFast
 from formerbox.utils.code_tokenizer import SpecialToken
 from tokenizers import AddedToken
 from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
 Token = Union[Text, AddedToken]
 TransformersTokenizer = Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
+
+logger = logging.getLogger(__name__)
 
 
 VOCAB_FILES_NAMES = {
@@ -22,7 +25,7 @@ PRETRAINED_VOCAB_FILES_MAP = {
 PRETRAINED_POSITIONAL_EMBEDDINGS_SIZES = {}
 
 
-class CodeTokenizerFast(TransformerTokenizerFast):
+class CodeBBPETokenizerFast(ByteLevelBPETokenizerFast):
     vocab_files_names = VOCAB_FILES_NAMES
     pretrained_vocab_files_map = PRETRAINED_VOCAB_FILES_MAP
     max_model_input_sizes = PRETRAINED_POSITIONAL_EMBEDDINGS_SIZES
@@ -32,6 +35,8 @@ class CodeTokenizerFast(TransformerTokenizerFast):
         self,
         vocab_file: Text,
         merges_file: Text,
+        tokenizer_file: Optional[Text] = None,
+        errors: Text = "replace",
         unk_token: Token = "<unk>",
         bos_token: Token = "<s>",
         eos_token: Token = "</s>",
@@ -46,30 +51,29 @@ class CodeTokenizerFast(TransformerTokenizerFast):
         **kwargs: Any,
     ) -> None:
         # pylint: disable=too-many-locals, too-many-arguments
-
-        # add default special tokens
-        kwargs.setdefault("unk_token", unk_token)
-        kwargs.setdefault("bos_token", bos_token)
-        kwargs.setdefault("eos_token", eos_token)
-        kwargs.setdefault("pad_token", pad_token)
-        kwargs.setdefault("sep_token", sep_token)
-        kwargs.setdefault("cls_token", cls_token)
-        kwargs.setdefault("mask_token", mask_token)
-
-        # add code specific special tokens
-        additional_special_tokens: List[Token] = []
-        for token in SpecialToken:
-            kwargs.setdefault(token.name, token.value)
-            additional_special_tokens.append(token.value)
-
-        kwargs.setdefault("additional_special_tokens", additional_special_tokens)
-
         super().__init__(
-            vocab_file=vocab_file,
-            merges_file=merges_file,
+            vocab_file,
+            merges_file,
+            tokenizer_file=tokenizer_file,
+            errors=errors,
+            bos_token=bos_token,
+            eos_token=eos_token,
+            sep_token=sep_token,
+            cls_token=cls_token,
+            unk_token=unk_token,
+            pad_token=pad_token,
+            mask_token=mask_token,
             add_prefix_space=add_prefix_space,
             trim_offsets=trim_offsets,
             lowercase=lowercase,
             dropout=dropout,
             **kwargs,
         )
+
+        # add code specific special tokens
+        additional_special_tokens: List[Text] = []
+        for token in SpecialToken:
+            setattr(self, token.name, token.value)
+            additional_special_tokens.append(token.value)
+
+        self.additional_special_tokens += additional_special_tokens