+++ formerbox/tasks/code/code_tokenizer_module.py
@@ -1,9 +1,10 @@
+import logging
 from pathlib import Path
 from typing import Any, Text, Union
 
 from formerbox.modules import TokenizerModule
-from formerbox.tasks import TransformerTokenizerModule
-from formerbox.tasks.code.code_tokenization import CodeTokenizerFast
+from formerbox.tasks import ByteLevelBPETokenizerModule
+from formerbox.tasks.code.code_tokenization import CodeBBPETokenizerFast
 from formerbox.utils.code_tokenizer import SpecialToken
 from tokenizers import AddedToken
 from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
@@ -11,10 +12,12 @@ from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 Token = Union[Text, AddedToken]
 TransformersTokenizer = Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
 
+logger = logging.getLogger(__name__)
 
-@TokenizerModule.register("code-tokenizer-fast", constructor="from_partial")
-class CodeTokenizerModule(TransformerTokenizerModule):
-    Params = TransformerTokenizerModule.Params
+
+@TokenizerModule.register("code-bbpe-tokenizer", constructor="from_partial")
+class CodeBBPETokenizerModule(ByteLevelBPETokenizerModule):
+    Params = ByteLevelBPETokenizerModule.Params
 
     def __init__(self, params: Params, **kwargs: Any) -> None:
         super().__init__(params, **kwargs)
@@ -30,7 +33,7 @@ class CodeTokenizerModule(TransformerTokenizerModule):
             tokenizer_path = Path(tokenizer_path)
         vocab_file = str(tokenizer_path / "vocab.json")
         merges_file = str(tokenizer_path / "merges.txt")
-        return CodeTokenizerFast(
+        return CodeBBPETokenizerFast(
             vocab_file=vocab_file, merges_file=merges_file, **kwargs
         )
 
@@ -38,11 +41,13 @@ class CodeTokenizerModule(TransformerTokenizerModule):
     def from_pretrained(cls, params: Params, **kwargs: Any) -> TransformersTokenizer:
         # prepare init arguments from params
         assert params.tokenizer_path is not None
-        init_kwargs = cls.get_args(params)
+        init_kwargs = cls.get_tokenizer_args(params)
         kwargs.update(init_kwargs)
 
         # get the pretrained tokenizer
-        tokenizer = CodeTokenizerFast.from_pretrained(params.tokenizer_path, **kwargs)
+        tokenizer = CodeBBPETokenizerFast.from_pretrained(
+            params.tokenizer_path, **kwargs
+        )
         assert isinstance(tokenizer, PreTrainedTokenizerFast)
 
         return tokenizer