+++ formerbox/tasks/transformer_tokenizer_module.py
@@ -1,32 +1,87 @@
+import logging
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Text, Union
 
 from formerbox.common.dataclass_argparse import DataclassBase
-from formerbox.modules import TokenizerModule
-from formerbox.tasks.transformer_tokenization import TransformerTokenizerFast
+from formerbox.modules.tokenizer_module import ParamsType, TokenizerModule
+from formerbox.tasks.transformer_tokenization import ByteLevelBPETokenizerFast
 from formerbox.utils.utils import path_to_posix
 from tokenizers import AddedToken
+from tokenizers.implementations import BaseTokenizer as FastTokenizer
 from tokenizers.implementations import ByteLevelBPETokenizer
 from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
 Token = Union[Text, AddedToken]
 TransformersTokenizer = Union[PreTrainedTokenizer, PreTrainedTokenizerFast]
 
+logger = logging.getLogger(__name__)
 
-def fix_tokenizer(tokenizer: TransformersTokenizer) -> None:
-    init_kwargs = {}
-    for key, value in tokenizer.init_kwargs.items():
-        if isinstance(value, AddedToken):
-            init_kwargs[key] = str(value)
+
+SPECIAL_TOKENS: List[Token] = [
+    "<s>",
+    "<pad>",
+    "</s>",
+    "<unk>",
+    "<mask>",
+]
+
+
+class TransformerTokenizerModule(TokenizerModule[ParamsType]):
+    special_tokens: List[Token]
+    tokenizer: FastTokenizer
+
+    def __init__(self, params: ParamsType, **kwargs: Any) -> None:
+        super().__init__(**kwargs)
+        self.params = params
+        self.special_tokens: List[Token] = SPECIAL_TOKENS
+        self.tokenizer = self.build_tokenizer(params)
+
+    @classmethod
+    def build_tokenizer(cls, params: ParamsType) -> FastTokenizer:
+        raise NotImplementedError()
+
+    @classmethod
+    def get_tokenizer_args(cls, params: ParamsType) -> Dict[Text, Any]:
+        raise NotImplementedError()
+
+    def fix_tokenizer(self, tokenizer: TransformersTokenizer) -> None:
+        init_kwargs = getattr(tokenizer, "init_kwargs", {})
+        for key, value in init_kwargs.items():
+            if isinstance(value, AddedToken):
+                init_kwargs[key] = str(value)
+            else:
+                init_kwargs[key] = value
+
+    def save_pretrained(
+        self, save_directory: Text, legacy_format: bool, **kwargs: Any
+    ) -> None:
+        # make sure the `tokenizer_output_path` is a pathlike object
+        if isinstance(save_directory, str):
+            tokenizer_path = Path(save_directory)
         else:
-            init_kwargs[key] = value
+            tokenizer_path = save_directory
 
-    tokenizer.init_kwargs = init_kwargs
+        # make the output dir if it doesn't exist
+        tokenizer_path.mkdir(parents=True, exist_ok=True)
+
+        # save the trained tokenizer to `tokenizer_output_path`
+        self.tokenizer.save_model(save_directory)
+
+        # prepare the pre-trained tokenizer
+        tokenizer = self.configure_tokenizer(tokenizer_path=tokenizer_path, **kwargs)
+
+        # workaround for saving tokenizer bugs in the transformers backend
+        self.fix_tokenizer(tokenizer)
+
+        # save the pre-trained tokenizer
+        tokenizer.save_pretrained(
+            save_directory=save_directory, legacy_format=legacy_format
+        )
 
 
-@TokenizerModule.register(name="transformer-tokenizer-fast", constructor="from_partial")
-class TransformerTokenizerModule(TokenizerModule):
+@TokenizerModule.register(name="byte-level-bpe-tokenizer", constructor="from_partial")
+class ByteLevelBPETokenizerModule(TransformerTokenizerModule):
     # pylint: disable=arguments-differ
     @dataclass
     class Params(DataclassBase):
@@ -41,6 +96,14 @@ class TransformerTokenizerModule(TokenizerModule):
         min_frequency: int = field(
             default=2,
             metadata={"help": "The min frequency for calculating subwords merges."},
+        )
+        legacy_format: bool = field(
+            default=True,
+            metadata={
+                "help": "Whether to save the tokenizer in legacy format (default),"
+                " i.e. with tokenizer specific vocabulary and separate added_tokens files"
+                " in the unified JSON file format of the `tokenizers` library."
+            },
         )
         tokenizer_path: Optional[Text] = field(
             default=None,
@@ -92,17 +155,11 @@ class TransformerTokenizerModule(TokenizerModule):
     params_type = Params
 
     def __init__(self, params: Params, **kwargs: Any) -> None:
-        super().__init__(**kwargs)
-        self.params = params
-        self.special_tokens: List[Token] = [
-            "<s>",
-            "<pad>",
-            "</s>",
-            "<unk>",
-            "<mask>",
-        ]
+        super().__init__(params, **kwargs)
 
-        self.backend_tokenizer = ByteLevelBPETokenizer(
+    @classmethod
+    def build_tokenizer(cls, params: Params) -> ByteLevelBPETokenizer:
+        return ByteLevelBPETokenizer(
             add_prefix_space=params.add_prefix_space,
             lowercase=params.lowercase,
             dropout=params.dropout,
@@ -115,20 +172,29 @@ class TransformerTokenizerModule(TokenizerModule):
     def configure_tokenizer(
         self, tokenizer_path: Union[Text, Path], **kwargs: Any
     ) -> TransformersTokenizer:
+        # prepare paths for the tokenizer files
         if isinstance(tokenizer_path, str):
             tokenizer_path = Path(tokenizer_path)
+
         vocab_file = path_to_posix(tokenizer_path / "vocab.json")
         merges_file = path_to_posix(tokenizer_path / "merges.txt")
-        return TransformerTokenizerFast(
-            vocab_file=vocab_file, merges_file=merges_file, **kwargs
+
+        # configure the pretrained tokenizer
+        return ByteLevelBPETokenizerFast(
+            vocab_file=vocab_file,
+            merges_file=merges_file,
+            **kwargs,
         )
 
     def train_tokenizer(self, *args: Any, **kwargs: Any) -> None:
         del args, kwargs  # use designated args
+
         # train a tokenizer model
         assert self.params.files is not None
         assert self.params.vocab_size is not None
-        self.backend_tokenizer.train(
+        assert isinstance(self.tokenizer, ByteLevelBPETokenizer)
+
+        self.tokenizer.train(
             files=self.params.files,
             vocab_size=self.params.vocab_size,
             min_frequency=self.params.min_frequency,
@@ -143,43 +209,33 @@ class TransformerTokenizerModule(TokenizerModule):
             assert self.params.tokenizer_path is not None
             save_directory = self.params.tokenizer_path
 
-        # make sure the `tokenizer_output_path` is a pathlike object
-        if isinstance(save_directory, str):
-            tokenizer_path = Path(save_directory)
-        else:
-            tokenizer_path = save_directory
+        # get the `legacy_format` argument value
+        legacy_format = kwargs.pop("legacy_format", self.params.legacy_format)
 
-        # make the output dir if it doesn't exist
-        tokenizer_path.mkdir(exist_ok=True)
-
-        # save the trained tokenizer to `tokenizer_output_path`
-        self.backend_tokenizer.save_model(save_directory)
-
-        # prepare the pre-trained tokenizer
-        tokenizer = self.configure_tokenizer(tokenizer_path=tokenizer_path, **kwargs)
-
-        # workaround for saving tokenizer bugs in the transformers backend
-        fix_tokenizer(tokenizer)
-        # save the pre-trained tokenizer
-        tokenizer.save_pretrained(save_directory)
+        # save the pretrained tokenizer to `save_directory`
+        super().save_pretrained(
+            save_directory=save_directory,
+            legacy_format=legacy_format,
+            **kwargs,
+        )
 
     @classmethod
     def from_pretrained(cls, params: Params, **kwargs: Any) -> TransformersTokenizer:
         # prepare init arguments from params
         assert params.tokenizer_path is not None
-        init_kwargs = cls.get_args(params)
+        init_kwargs = cls.get_tokenizer_args(params)
         kwargs.update(init_kwargs)
 
         # get the pretrained tokenizer
-        tokenizer = TransformerTokenizerFast.from_pretrained(
+        tokenizer = ByteLevelBPETokenizerFast.from_pretrained(
             params.tokenizer_path, **kwargs
         )
         assert isinstance(tokenizer, PreTrainedTokenizerFast)
 
         return tokenizer
 
-    @staticmethod
-    def get_args(params: Params) -> Dict[Text, Any]:
+    @classmethod
+    def get_tokenizer_args(cls, params: Params) -> Dict[Text, Any]:
         # prepare a copy of args for the pretrained tokenizer
         kwargs = vars(params).copy()